Added Mocking recipe and general setup for recipies
