+++ docs/.vitepress/routes/sidebar/index.ts
@@ -1,0 +1,9 @@
+import type { DefaultTheme } from 'vitepress'
+
+import { routes as guideRoutes } from './guide'
+import { routes as recipesRoutes } from './recipes'
+
+export const routes: DefaultTheme.Config['sidebar'] = {
+  '/guide': guideRoutes,
+  '/recipes': recipesRoutes
+}
