+++ docs/.vitepress/routes/sidebar/recipes.ts
@@ -1,0 +1,33 @@
+import type { DefaultTheme } from 'vitepress'
+
+export const routes: DefaultTheme.SidebarItem[] = [
+  {
+    text: 'Introduction',
+    base: '/recipes/introduction',
+    items: [
+      {
+        text: 'Welcome',
+        link: '/welcome',
+      },
+      {
+        text: 'Adding your recipe',
+        link: '/adding-your-recipe',
+      }
+    ],
+  },
+  {
+    text: 'Official',
+    base: '/recipes/official',
+    items: [
+      {
+        text: 'Mocking with Vitest',
+        link: '/mocking-with-vitest',
+      },
+    ],
+  },
+  {
+    text: 'Community',
+    base: '/recipes/community',
+    items: [],
+  },
+]
