+++ docs/recipes/introduction/adding-your-recipe.md
@@ -1,0 +1,16 @@
+# Adding your recipes
+
+We appreciate anyone willing to share their recipes with the community! Authentication is a complex requirement, individual to every project, therefore the more experiences the community shares with one another the better!
+
+We are willing to add recipes that come in different media forms. These could be:
+
+- A blog post
+- A step-by-step guide
+- A GitHub repository
+- _Any other media you prefer!_
+
+However, we have noticed that the more detailed a recipe is the more helpful it can be to the community!
+
+## Submitting your recipe
+
+If you would like to submit a recipe, please open a Pull request, with your recipe added to `~/docs/recipes/community`.
