+++ docs/recipes/introduction/welcome.md
@@ -1,0 +1,13 @@
+# Recipes
+
+The following pages contain recipes for commonly asked patterns, questions, and implementations. The recipes are mostly provided by the community and can serve as guidelines to implement something similar in your Nuxt 3 application.
+
+## What are recipes?
+
+Recipes are short guides provided by the community and our team that outline how to use `@sidebase/nuxt-auth` in specific use cases required by your application. 
+
+Recipes are not meant to act as official documentation and should be primarily used as inspiration to customize `@sidebase/nuxt-auth` to your needs. 
+
+:::warning
+The recipes are not all tested through by the sidebase team. If you have any concerns, questions or improvement proposals or want to contribute a recipe yourself, we'd be very happy if you [open an issue on our repository](https://github.com/sidebase/nuxt-auth/issues/new/choose).
+:::
