+++ docs/recipes/official/overview.md
@@ -1,0 +1,3 @@
+# Coming soon
+
+Oops, this page is not ready yet. You can find our current docs [here](https://sidebase.io).
